♻️ 63FTP6 refactor: reduce duplicate paths
