+++ .agentplane/tasks/202605251947-63FTP6/pr/meta.json
@@ -1,0 +1,13 @@
+{
+  "base": "main",
+  "branch": "task/202605251947-63FTP6/reduce-duplicate-paths",
+  "created_at": "2026-05-25T19:51:43.995Z",
+  "diffstat_sha256": "sha256:cc83efa90e33c073bb349b8132888359082d8dee7eadb21f456c80743b3cbcb5",
+  "last_verified_at": null,
+  "schema_version": 1,
+  "task_id": "202605251947-63FTP6",
+  "updated_at": "2026-05-25T19:51:43.995Z",
+  "verify": {
+    "status": "skipped"
+  }
+}
