+++ .agentplane/tasks/202605251947-63FTP6/pr/review.md
@@ -1,0 +1,40 @@
+# PR Review
+
+Created: 2026-05-25T19:51:43.995Z
+
+## Task
+
+- Task: `202605251947-63FTP6`
+- Title: Reduce low-risk duplicate implementation paths
+- Status: DOING
+- Branch: `task/202605251947-63FTP6/reduce-duplicate-paths`
+- Canonical task record: `.agentplane/tasks/202605251947-63FTP6/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-25T19:51:43.995Z
+- Branch: task/202605251947-63FTP6/reduce-duplicate-paths
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../agentplane/src/commands/task/verify-record.ts  | 102 +++++++++------------
+ scripts/lib/workflow-config.mjs                    |  64 +++++++++++++
+ scripts/release/release-task-evidence.mjs          |  55 +----------
+ scripts/workflow/prepare-hosted-task-closure.mjs   |  59 +-----------
+ 4 files changed, 112 insertions(+), 168 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
