+++ packages/agentplane/src/commands/task/verify-record.ts
@@ -1,6 +1,48 @@
 import type { CommandContext } from "../shared/task-backend.js";
 
 import { executeVerifyRecordCommand } from "./verify-record-execute.js";
+import type { VerifyStructuredFindingInput } from "./verify-record.types.js";
+
+type VerifyFindingOptions = {
+  observation?: string;
+  impact?: string;
+  resolution?: string;
+  promote?: boolean;
+  external?: boolean;
+  localOnly?: boolean;
+  repoFixable?: boolean;
+  incidentScope?: string;
+  incidentTags?: string[];
+  incidentMatch?: string[];
+  incidentAdvice?: string;
+  incidentRule?: string;
+};
+
+function buildVerifyStructuredFinding(
+  opts: VerifyFindingOptions,
+): VerifyStructuredFindingInput | null {
+  if (
+    typeof opts.observation !== "string" ||
+    typeof opts.impact !== "string" ||
+    typeof opts.resolution !== "string"
+  ) {
+    return null;
+  }
+  return {
+    observation: opts.observation,
+    impact: opts.impact,
+    resolution: opts.resolution,
+    promote: opts.promote === true,
+    external: opts.external === true,
+    localOnly: opts.localOnly === true,
+    repoFixable: opts.repoFixable === true,
+    incidentScope: opts.incidentScope,
+    incidentTags: opts.incidentTags ?? [],
+    incidentMatch: opts.incidentMatch ?? [],
+    incidentAdvice: opts.incidentAdvice,
+    incidentRule: opts.incidentRule,
+  };
+}
 
 export async function cmdTaskVerifyOk(opts: {
   ctx?: CommandContext;
@@ -31,25 +73,7 @@ export async function cmdTaskVerifyOk(opts: {
     ...opts,
     state: "ok",
     command: "task verify ok",
-    finding:
-      typeof opts.observation === "string" &&
-      typeof opts.impact === "string" &&
-      typeof opts.resolution === "string"
-        ? {
-            observation: opts.observation,
-            impact: opts.impact,
-            resolution: opts.resolution,
-            promote: opts.promote === true,
-            external: opts.external === true,
-            localOnly: opts.localOnly === true,
-            repoFixable: opts.repoFixable === true,
-            incidentScope: opts.incidentScope,
-            incidentTags: opts.incidentTags ?? [],
-            incidentMatch: opts.incidentMatch ?? [],
-            incidentAdvice: opts.incidentAdvice,
-            incidentRule: opts.incidentRule,
-          }
-        : null,
+    finding: buildVerifyStructuredFinding(opts),
   });
 }
 
@@ -82,25 +106,7 @@ export async function cmdTaskVerifyRework(opts: {
     ...opts,
     state: "needs_rework",
     command: "task verify rework",
-    finding:
-      typeof opts.observation === "string" &&
-      typeof opts.impact === "string" &&
-      typeof opts.resolution === "string"
-        ? {
-            observation: opts.observation,
-            impact: opts.impact,
-            resolution: opts.resolution,
-            promote: opts.promote === true,
-            external: opts.external === true,
-            localOnly: opts.localOnly === true,
-            repoFixable: opts.repoFixable === true,
-            incidentScope: opts.incidentScope,
-            incidentTags: opts.incidentTags ?? [],
-            incidentMatch: opts.incidentMatch ?? [],
-            incidentAdvice: opts.incidentAdvice,
-            incidentRule: opts.incidentRule,
-          }
-        : null,
+    finding: buildVerifyStructuredFinding(opts),
   });
 }
 
@@ -133,24 +139,6 @@ export async function cmdVerifyParsed(opts: {
   return await executeVerifyRecordCommand({
     ...opts,
     command: "verify",
-    finding:
-      typeof opts.observation === "string" &&
-      typeof opts.impact === "string" &&
-      typeof opts.resolution === "string"
-        ? {
-            observation: opts.observation,
-            impact: opts.impact,
-            resolution: opts.resolution,
-            promote: opts.promote === true,
-            external: opts.external === true,
-            localOnly: opts.localOnly === true,
-            repoFixable: opts.repoFixable === true,
-            incidentScope: opts.incidentScope,
-            incidentTags: opts.incidentTags ?? [],
-            incidentMatch: opts.incidentMatch ?? [],
-            incidentAdvice: opts.incidentAdvice,
-            incidentRule: opts.incidentRule,
-          }
-        : null,
+    finding: buildVerifyStructuredFinding(opts),
   });
 }
