+++ scripts/release/release-task-evidence.mjs
@@ -11,6 +11,7 @@ import {
   taskReadmeDocBody,
   taskReadmePath,
 } from "../../packages/core/src/index.ts";
+import { readConfiguredTaskCloseBranchPrefix } from "../lib/workflow-config.mjs";
 
 const execFileAsync = promisify(execFile);
 
@@ -222,60 +223,6 @@ function parsePrepareArgs(argv) {
   out.publishResultPath = ensureNonEmpty(out.publishResultPath, "publish-result path");
   out.repo = ensureNonEmpty(out.repo, "repo");
   return out;
-}
-
-async function readConfiguredTaskCloseBranchPrefix(cwd = process.cwd()) {
-  try {
-    const workflow = await readFile(path.join(cwd, ".agentplane", "WORKFLOW.md"), "utf8");
-    const branch = parseWorkflowBranchConfig(workflow);
-    return stringValue(branch.task_close_prefix);
-  } catch {
-    return null;
-  }
-}
-
-function parseWorkflowBranchConfig(text) {
-  const normalized = String(text ?? "")
-    .replaceAll("\r\n", "\n")
-    .replaceAll("\r", "\n");
-  if (!normalized.startsWith("---\n")) return {};
-  const end = normalized.indexOf("\n---", 4);
-  if (end === -1) return {};
-  const out = {};
-  let inBranch = false;
-  for (const line of normalized.slice(4, end).split("\n")) {
-    if (line === "branch:") {
-      inBranch = true;
-      continue;
-    }
-    if (!inBranch) continue;
-    if (line.length > 0 && !line.startsWith("  ")) break;
-    const trimmed = line.trim();
-    const separator = trimmed.indexOf(":");
-    if (separator === -1) continue;
-    const key = trimmed.slice(0, separator).trim();
-    if (key !== "task_prefix" && key !== "task_close_prefix") continue;
-    out[key] = parseYamlStringScalar(trimmed.slice(separator + 1).trim());
-  }
-  return out;
-}
-
-function parseYamlStringScalar(value) {
-  if (!value) return null;
-  if (value.startsWith('"') && value.endsWith('"')) {
-    return value
-      .slice(1, -1)
-      .replaceAll(String.raw`\"`, '"')
-      .trim();
-  }
-  if (value.startsWith("'") && value.endsWith("'")) {
-    return value.slice(1, -1).replaceAll("''", "'").trim();
-  }
-  return value.split(" #", 1)[0]?.trim() ?? null;
-}
-
-function stringValue(value) {
-  return typeof value === "string" && value.trim().length > 0 ? value.trim() : null;
 }
 
 function parseApplyArgs(argv) {
